+++ lib/market_hub/api/mercado_livre/shipment.rb
@@ -11,7 +11,7 @@ module MarketHub
 
         def initialize(access_token)
           @access_token = access_token
-          @types = %w(items costs payments lead_time delays history carrier)
+          @types = %w(items costs payments lead_time delays history carrier billing_info)
         end
 
         def find(shipment_id, type = nil)
@@ -31,12 +31,12 @@ module MarketHub
           endpoint = URI::HTTPS.build(host: host, path: path)
           endpoint.query = URI.encode_www_form(params)
           response = MarketHub::HTTP.get(endpoint, headers: { authorization: "Bearer #{@access_token}" })
-          convert_to_json_or_file(response.body)
+          try_to_convert_to_json(response.body)
         end
 
         private
 
-          def convert_to_json_or_file(body)
+          def try_to_convert_to_json(body)
             JSON.parse(body)
           rescue
             body